chore: moved to cargo workspaces part A
