+++ Cargo.toml
@@ -1,66 +1,3 @@
-[package]
-name = "minecraft-essentials"
-version = "0.2.12"
-keywords = ["Minecraft", "MinecraftModpack", "Essentials"]
-authors = ["Eveeifyeve <eveeg1971@gmail.com>"]
-description = "A Library that gives all Minecraft Modpack developers essentials."
-homepage = "https://minecraft-essentials.github.io"
-repository = "https://github.com/minecraft-essentials/minecraft-essentials"
-documentation = "https://docs.rs/minecraft-essentials"
-license-file = "LICENSE"
-edition = "2021"
 
-[build]
-rustdocflags = ["--theme=ayu"]
-
-[[bin]]
-name = "minecraft-essentials"
-path = "src/main.rs"
-
-[package.metadata.binstall]
-pkg-url = "{ repo }/releases/download/v{ version }/{ name }-cli"
-bin-dir = "{ name }-{ target }-v{ version }/{ bin }{ binary-ext }"
-pkg-fmt = "tgz"
-
-[package.metadata.binstall.overrides.x86_64-pc-windows-msvc]
-pkg-fmt = "zip"
-
-[package.metadata.binstall.overrides.aarch64-pc-windows-msvc]
-pkg-fmt = "zip"
-
-[dependencies]
-tokio = { version = "1.37.0", features = ["rt-multi-thread", "macros", "sync"] }
-reqwest = { version = "0.12.4", features = ["json"], optional = true }
-serde = { version = "1.0.199", features = ["derive"], optional = true }
-clap = { version = "4.5.4", optional = true, features = ["derive"] }
-serde_urlencoded = { version = "0.7.1", optional = true }
-serde_json = { version = "1.0.116", optional = true }
-base64 = { version = "0.22.0", optional = true }
-ring = { version = "0.17.8", optional = true }
-rand = { version = "0.8.5", optional = true }
-hex = { version = "0.4.3", optional = true }
-url = { version = "2.5.0", optional = true }
-displaydoc = "0.2.4"
-thiserror = "1.0.59"
-bytes = "1.6.0"
-sha-1 = "0.10.1"
-
-[dev-dependencies]
-dotenv = { version = "0.15.0" }
-
-[features]
-default = ["auth", "cli", "launch", "modrinth", "curseforge"]
-auth = [
-    "dep:serde_urlencoded",
-    "dep:serde_json",
-    "dep:reqwest",
-    "dep:serde",
-    "dep:rand",
-    "dep:url",
-]
-launch = ["dep:serde_json", "dep:serde", "dep:reqwest"]
-refresh = ["auth"]
-# Mod Platforms
-modrinth = ["dep:reqwest"]
-curseforge = []
-cli = ["dep:clap", "auth"]
+[workspace]
+members = ["crates/core" "crates/cli"]
