+++ crate/core/src/auth/microsoft.rs
