+++ crate/core/src/auth/xbox.rs
