+++ crate/core/src/launch/java.rs
