+++ crate/core/src/lib.rs
