+++ crate/core/src/structs.rs
